use "inventory" instead of "upload"
