+++ scripts/run_inventory_pdfs_pmid.py
@@ -1,14 +1,14 @@
 import argparse
 from dsst_etl import get_db_engine, logger
 from dsst_etl.db import get_db_session
-from dsst_etl.upload_pdfs_title_is_pmid import UploadPDFsTitleIsPMID
+from dsst_etl.inventory_pdfs_pmid import DocumentInventoryPMID
 
 def main():
     db_session = get_db_session(get_db_engine())
     
     try:
-        uploader = UploadPDFsTitleIsPMID(db_session)
-        uploader.run()
+        inventory = DocumentInventoryPMID(db_session)
+        inventory.run()
     except Exception as e:
         logger.error(f"An error occurred: {str(e)}")
     finally:
